Partial rewrite to better support parsing nested html
The previous version had alot of limitations, due to flattening the tree before rendering, which meant it became almost impossible to figure out which text content belongs to which block.
For example if there are two paragraph blocks, with inner tags, how do you know when each paragraph block ends, if they are at the same level?

This commit adds support for:
* Linebreaks (<br>).
* Paragraphs with some styled text and some unstyled text (for example some text inside <strong> tags and some just regular text).
* Bottom margins on tags, such as paragraphs, headings, etc.
* Links <a href="">.
* More text styling tags: <u>, <ins>, <b>, <strong>, <i>, <em>, <s>.
* More text styling css: 'text-decoration: line-through;'
* Bottom margin css from either 'margin' or 'margin-bottom'.
* Custom css styles through the optional options.style parameter.
* Override base fonts/fontsize through the optional options.fonts/options.fontSizes.base parameters.
* Override base color and link color through the optional options.colors.base/options.colors.link parameters.

Library is now no longer asyncronous, there seem to be no need for that anymore.
The transformations handling is also removed, as there seem to be no need for that either, now that the flattening of the html through events has been replaced with a tree parser instead.